ml/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
nltk==3.5
numpy==1.19.4
pandas==1.1.4
protobuf==3.14.0
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.54.0
Werkzeug==1.0.1
